fix(review): move path in const file
