+++ packages/client/src/Pages/Forum/lib/mocks.ts
@@ -1,0 +1,54 @@
+import { Topics } from '@models/types/topics';
+
+export const TOPICS_LIST: Topics = [
+	{
+		id: 0,
+		name: 'Interesting projects',
+		comments: 60
+	},
+	{
+		id: 1,
+		name: 'Secret bonuses',
+		comments: 100
+	},
+	{
+		id: 2,
+		name: 'Other theme',
+		comments: 2244
+	},
+	{
+		id: 3,
+		name: 'Topic 1',
+		comments: 1
+	},
+	{
+		id: 4,
+		name: 'Chits',
+		comments: 121300842394
+	},
+	{
+		id: 5,
+		name: 'Long name for test naming and some more words and phrases and more and more and more',
+		comments: 100000
+	},
+	{
+		id: 6,
+		name: 'Topic 2',
+		comments: 60
+	},
+	{
+		id: 7,
+		name: 'Topic 10',
+		comments: 60
+	},
+	{
+		id: 8,
+		name: 'No name',
+		comments: 60
+	},
+	{
+		id: 9,
+		name: 'DEFAULT',
+		comments: 60
+	}
+];
